Renamed LoginScreen package to menueGui
