+++ src/menueGui/Button.java
@@ -1,4 +1,4 @@
-package LoginScreen;
+package menueGui;
 
 /*
  * written by Elias Geiger and Ben Brandes
